Refactor and reorganize ComponentEditor hooks
