+++ react/components/EditorContainer/Sidebar/ComponentEditor/typings.d.ts
@@ -1,5 +1,11 @@
 import { JSONSchema6 } from 'json-schema'
 
+export interface ComponentFormState {
+  onClose: () => void
+  onTitleChange?: () => void
+  title: string
+}
+
 export interface GetSchemasArgs {
   contentSchema?: JSONSchema6
   editTreePath: string | null
